Исправление парсинга головы сообщения, улучшения работы с буфером, мелкие исправления логгирования. Добавлены еще тесты.

+++ SpectrumVisualizer.Uart/SpectrumJobs/DataStruct.cs
@@ -2,7 +2,7 @@
 {
     public class DataStruct(int spectrumLength)
     {
-        public readonly double[] Spectrum = new double[spectrumLength];
+        public readonly ushort[] Spectrum = new ushort[spectrumLength];
 
         public double Average;
